requirements.txt
telegram
python-telegram-bot
pandas
pillow
